# Minor Project/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/admin/details": {
            "get": {
                "tags": [
                    "admin details"
                ],
                "summary": "admin details",
                "operationId": "details",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/user/form": {
            "post": {
                "tags": [
                    "submit user form"
                ],
                "summary": "submit the user form for a particular company",
                "operationId": "submit_form",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "description": "Updated name of the pet",
                                        "type": "string"
                                    },
                                    "job_id": {
                                        "description": "Updated name of the pet",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Updated name of the pet",
                                        "type": "string"
                                    },
                                    "pdf": {
                                        "description": "Updated name of the pet",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/update_user_form/{job_id}": {
            "get": {
                "tags": [
                    "user form details by job id"
                ],
                "summary": "user form details of applied user for a job",
                "operationId": "user_form_display",
                "parameters": [
                    {
                        "name": "job_id",
                        "in": "path",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "default": "available"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/update_user_form/{id}": {
            "post": {
                "tags": [
                    "update user form with an id"
                ],
                "summary": "update intern",
                "operationId": "update_user_form",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "default": "available"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "description": "Updated name of the pet",
                                        "type": "string"
                                    },
                                    "job_id": {
                                        "description": "Updated name of the pet",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Updated name of the pet",
                                        "type": "string"
                                    },
                                    "pdf": {
                                        "description": "Updated name of the pet",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/jobs_posts": {
            "get": {
                "tags": [
                    "applied jobs by user"
                ],
                "summary": "get all the job post where user has applied intern for",
                "operationId": "get_all_applied_job_forms",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/delete_user_form/{id}": {
            "post": {
                "tags": [
                    "delete user form"
                ],
                "summary": "delete user form",
                "operationId": "delete_user_form",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user form id",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "default": "available"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/company/job_posts/{job_id}": {
            "post": {
                "tags": [
                    "get all user form"
                ],
                "summary": "get all user form for a given job post",
                "operationId": "get_applied_user_form",
                "parameters": [
                    {
                        "name": "job_id",
                        "in": "path",
                        "description": "job post id",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "default": "available"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/company/job_posts/{job_id}/{id}": {
            "post": {
                "tags": [
                    "get user form pdf"
                ],
                "summary": "get user form  pdf for a given job post",
                "operationId": "user_form_pdf",
                "parameters": [
                    {
                        "name": "job_id",
                        "in": "path",
                        "description": "job post id",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "default": "available"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user form id",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "default": "available"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/company/job_post/{job_id}/{id}": {
            "post": {
                "tags": [
                    "get user details"
                ],
                "summary": "get user form  details who has applied for a given job post",
                "operationId": "user_details",
                "parameters": [
                    {
                        "name": "job_id",
                        "in": "path",
                        "description": "job post id",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "default": "available"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "default": "available"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}